requirements.txt
bottle
python-dotenv
requests